turns out currentTimeMillis() method  is type long

+++ Assignment_#2/countStringInText.java
@@ -116,13 +116,13 @@ public class countStringInText {
         LinkedList<> patternForLinkedList = makeLinkedListStr(pattern);
         ArrayList<> patternForArrayList = makeArrayListStr(pattern);
 
-        int originalTime = System.currentTimeMillis(); //current time, thanks for the hint prof haha
+        long originalTime = System.currentTimeMillis(); //current time, thanks for the hint prof haha
         
         int countLinkedList = searchListSpecificWord(filename, patternForLinkedList);
-        int elapsedTimeLinkedList = System.currentTimeMillis() - originalTime; //elapsed time for linked list
+        long elapsedTimeLinkedList = System.currentTimeMillis() - originalTime; //elapsed time for linked list
 
         int countArrayList = searchListSpecificWord(filename, patternForArrayList);
-        int elapsedTimeArrayList = System.currentTimeMillis() - originalTime;
+        long elapsedTimeArrayList = System.currentTimeMillis() - originalTime;
 
     }
 }
